+++ src/robot_impl.cpp
@@ -256,7 +256,7 @@ uint32_t Robot::Impl::startMotion(
     const research_interface::robot::Move::Deviation& maximum_path_deviation,
     const research_interface::robot::Move::Deviation& maximum_goal_pose_deviation,
     bool use_async_motion_generator,
-    const std::optional<std::array<double, kNumJoints>>& maximum_velocities) {
+    const std::optional<std::vector<double>>& maximum_velocities) {
   if (motionGeneratorRunning() || controllerRunning()) {
     throw ControlException("libfranka robot: Attempted to start multiple motions!");
   }