+++ CHANGELOG.md
@@ -7,12 +7,18 @@ All notable changes to libfranka in this file.
 #### Changed
 - Hotfix to avoid torques discontinuity false positives due to robot state float precision change.
 
+### pylibfranka - Python
+#### Added
+- Async control python bindings
+- Async joint positions control example from C++.
+
 ## [0.19.0]
 ### libfranka - C++
 #### Changed
 - To support franka_ros2, we added an option for the async position control to base the `getFeedback` function on a robot state received via `franka_hardware` instead of querying the robot directly.
 - Format libfranka debian package to inlclude ubuntu code name and arch: libfranka_VERSION_CODENAME_ARCH.deb
 - Added build containers to support Ubuntu 22.04 and 24.04
+
 
 ## [0.18.2]
 Requires Franka Research 3 System Version >= 5.9.0